Move App to `/scenes/`.
